Separate vehicle counter class from execution loop
